Improving structure
After this commit, common utility functions will be added at `utilities`
module. All the solution programs can be dependent on this utility to
save time and move forward.
