+++ programs/tests/test_arrays/test_maximum_in_struct_array.py
@@ -7,34 +7,6 @@ from arrays import maximum_in_struct_array
 
 class TestMaximumInStructArray(unittest.TestCase):
 
-    def test_validate_for_invalid_values(self):
-        def _validate(value):
-            return ((value >= 100) and (value <= 200))
-
-        values_and_callables = (
-            ("Y", lambda c: len(c) > 2),
-            (2333, lambda w: (w >= 1) and (w <= 100)),
-            (400, _validate)
-        )
-        for value, _callable in values_and_callables:
-            with self.subTest(value=value, callable=_callable):
-                with self.assertRaises(ValueError):
-                    maximum_in_struct_array.validate(value, _callable)
-
-    def test_validate_for_correct_values(self):
-        def _validate(value):
-            return ((value >= 1) and (value <= 100))
-
-        values_and_callables = (
-            ("Y", lambda c: len(c) == 1),
-            (234, lambda n: (n >= 0) and (n <= 300)),
-            (20, _validate)
-        )
-
-        for value, _callable in values_and_callables:
-            with self.subTest(value=value, callable=_callable):
-                maximum_in_struct_array.validate(value, _callable)
-
     def test_inches(self):
         values_and_expected_return_values = (
             (2, 24),
